unit-file: simplify & make more reliable
When some `Exec*` statements fail, systemd should be aware of it, and
mark the service as failed. Those see their `=-` (which means lax exit
code checking) changed to the stricter `=`.
The other `Exec*` statements keep their `=-` as they are not critical,
but can be simplified, as their exit status doesn't matter.